requirements.txt
dj-database-url==0.4.1
django-pylibmc==0.6.1
django-simple-history>=3.0.0
Django==4.2
gunicorn>=21.0.0
Markdown>=3.0
psycopg2>=2.9.9
pylibmc>=1.6.0
whitenoise>=6.0.0
requests>=2.31.0
